# internal/auth-service/v1/service/handlers.go
package service

import (
	"encoding/json"
	"net/http"

	"go.uber.org/zap"
)

type handlers struct {
	logger *zap.SugaredLogger
}

func New(logger *zap.SugaredLogger) *handlers {
	return &handlers{
		logger: logger,
	}
}

// @Summary Register a new user
// @Description creates a new user account using email and password
// @Tags Auth
// @Accent json
// @Produce json
// @Param request body map[string]string true "User credentials"
// @Success 201 {object} map[string]string
// @Failure 400 {object} map[string]string
// @Router /auth/register [post]
func (handlers) PostAuthRegister(w http.ResponseWriter, r *http.Request) {
	json.NewEncoder(w).Encode(map[string]string{"message": "Register endpoint"})
}

// @Summary Login user
// @Description Authenticates a user and returns access & refresh tokens
// @Tags Auth
// @Accept  json
// @Produce  json
// @Param request body map[string]string true "Login credentials"
// @Success 200 {object} map[string]string
// @Failure 401 {object} map[string]string
// @Router /auth/login [post]
func (handlers) PostAuthLogin(w http.ResponseWriter, r *http.Request) {
	json.NewEncoder(w).Encode(map[string]string{"message": "Login endpoint"})
}

// @Summary Refresh access token
// @Description Issues new access token using refresh token
// @Tags Auth
// @Accept  json
// @Produce  json
// @Param request body map[string]string true "Refresh token request"
// @Success 200 {object} map[string]string
// @Router /auth/refresh [post]
func (handlers) PostAuthRefresh(w http.ResponseWriter, r *http.Request) {
	json.NewEncoder(w).Encode(map[string]string{"message": "Refresh endpoint"})
}

// @Summary Logout user
// @Description Invalidates active token/session
// @Tags Auth
// @Security BearerAuth
// @Produce  json
// @Success 200 {object} map[string]string
// @Router /auth/logout [post]
func (handlers) PostAuthLogout(w http.ResponseWriter, r *http.Request) {
	json.NewEncoder(w).Encode(map[string]string{"message": "Logout endpoint"})
}

// @Summary Get current user profile
// @Description Returns details of authenticated user
// @Tags Auth
// @Security BearerAuth
// @Produce  json
// @Success 200 {object} map[string]string
// @Router /auth/me [get]
func (handlers) GetAuthMe(w http.ResponseWriter, r *http.Request) {
	json.NewEncoder(w).Encode(map[string]string{"message": "Profile endpoint"})
}

// @Summary Change password
// @Description Allows logged-in user to change password
// @Tags Auth
// @Security BearerAuth
// @Accept  json
// @Produce  json
// @Param request body map[string]string true "Password change request"
// @Success 200 {object} map[string]string
// @Router /auth/change-password [post]
func (handlers) PostAuthChangePassword(w http.ResponseWriter, r *http.Request) {
	json.NewEncoder(w).Encode(map[string]string{"message": "Change password endpoint"})
}

// @Summary Forgot password
// @Description Initiates password reset via email
// @Tags Auth
// @Accept  json
// @Produce  json
// @Param request body map[string]string true "Forgot password request"
// @Success 200 {object} map[string]string
// @Router /auth/forgot-password [post]
func (handlers) PostAuthForgotPassword(w http.ResponseWriter, r *http.Request) {
	json.NewEncoder(w).Encode(map[string]string{"message": "Forgot password endpoint"})
}

// @Summary Reset password
// @Description Completes password reset with new password
// @Tags Auth
// @Accept  json
// @Produce  json
// @Param request body map[string]string true "Reset password request"
// @Success 200 {object} map[string]string
// @Router /auth/reset-password [post]
func (handlers) PostAuthResetPassword(w http.ResponseWriter, r *http.Request) {
	json.NewEncoder(w).Encode(map[string]string{"message": "Reset password endpoint"})
}

// @Summary Health Check
// @Description Check if API Gateway is running
// @Tags Health
// @Produce json
// @Success 200 {object} map[string]string
// @Router /healthz [get]
func (handlers) GetHealthz(w http.ResponseWriter, r *http.Request) {
	json.NewEncoder(w).Encode(map[string]string{"status": "ok"})
}